fix de personnalisation et du mm cas du warning du tableau temporaire

+++ src/app/heavy-client/model/simulation-engine/Arbre/Noeuds/NoeudRobot.h
@@ -101,7 +101,7 @@ public:
 
 	std::stack <NoeudAbstrait*> tableauCoins;
 
-	virtual void initialisationCouleurs(float* roues, float* modele);
+	virtual void initialisationCouleurs(int* roues, int* modele);
 
 
 protected:
